Create a function named playRound
Create a function named playRound where it will simulate a round of
rock, paper, scissors between the player and the computer. It will say
the string "It's a draw!" if the player and the computer's choices are
the same, "You win!, humanChoice, beats, computerChoice" if the player
wins, and "You lose!, computerChoice, beats, humanChoice" if the
computer beats the player. It will also make it so that the respective
scores of the player and the computer increases by 1 for every win
they obtain.